refactor(sweets): improve Sweet model, validation, and permissions for clean DRF implementation

+++ backend/sweets/models.py
@@ -3,8 +3,8 @@ from django.db import models
 class Sweet(models.Model):
     name = models.CharField(max_length=100, unique=True)
     category = models.CharField(max_length=100)
-    price = models.FloatField()
-    quantity = models.IntegerField()
+    price = models.DecimalField(max_digits=8, decimal_places=2)
+    quantity = models.PositiveIntegerField()
 
     def __str__(self):
         return self.name
